Revert "Add Parents and Subs to Keep Workplaces Inactive  (#3164)" (#3180)
This reverts commit cf8396a837becbe5b612c6b26e31df01913a6c0c.
