+++ server/test/unit/routes/admin/email-campaigns/inactive-workplaces/index.spec.js
@@ -1,11 +1,9 @@
 const expect = require('chai').expect;
 const httpMocks = require('node-mocks-http');
 const sinon = require('sinon');
-const moment = require('moment');
 
 const models = require('../../../../../../models');
 const findInactiveWorkplaces = require('../../../../../../services/email-campaigns/inactive-workplaces/findInactiveWorkplaces');
-const findParentWorkplaces = require('../../../../../../services/email-campaigns/inactive-workplaces/findParentWorkplaces');
 const sendEmail = require('../../../../../../services/email-campaigns/inactive-workplaces/sendEmail');
 const inactiveWorkplaceRoutes = require('../../../../../../routes/admin/email-campaigns/inactive-workplaces');
 
@@ -14,11 +12,6 @@ describe('server/routes/admin/email-campaigns/inactive-workplaces', () => {
     sinon.restore();
   });
 
-  const endOfLastMonth = moment().subtract(1, 'months').endOf('month').endOf('day');
-  const sixMonthTemplateId = 13;
-  const twelveMonthTemplateId = 14;
-  const parentTemplateId = 15;
-
   const dummyInactiveWorkplaces = [
     {
       id: 478,
@@ -26,7 +19,7 @@ describe('server/routes/admin/email-campaigns/inactive-workplaces', () => {
       nmdsId: 'J1234567',
       lastUpdated: '2020-06-01',
       emailTemplate: {
-        id: sixMonthTemplateId,
+        id: 13,
       },
       dataOwner: 'Workplace',
       user: {
@@ -40,7 +33,7 @@ describe('server/routes/admin/email-campaigns/inactive-workplaces', () => {
       nmdsId: 'A0012345',
       lastUpdated: '2020-01-01',
       emailTemplate: {
-        id: twelveMonthTemplateId,
+        id: 13,
       },
       dataOwner: 'Workplace',
       user: {
@@ -50,43 +43,8 @@ describe('server/routes/admin/email-campaigns/inactive-workplaces', () => {
     },
   ];
 
-  const dummyParentWorkplaces = [
-    {
-      id: 1,
-      name: 'Test Name',
-      nmdsId: 'A1234567',
-      lastUpdated: endOfLastMonth.clone().subtract(6, 'months').format('YYYY-MM-DD'),
-      emailTemplate: {
-        id: parentTemplateId,
-        name: 'Parent',
-      },
-      dataOwner: 'Workplace',
-      user: {
-        name: 'Test Person',
-        email: 'test@example.com',
-      },
-      subsidiaries: [
-        {
-          id: 2,
-          name: 'Workplace Name',
-          nmdsId: 'A0045232',
-          lastUpdated: endOfLastMonth.clone().subtract(6, 'months').format('YYYY-MM-DD'),
-          dataOwner: 'Parent',
-        },
-        {
-          id: 3,
-          name: 'Workplace Name',
-          nmdsId: 'A1245232',
-          lastUpdated: endOfLastMonth.clone().subtract(6, 'months').format('YYYY-MM-DD'),
-          dataOwner: 'Parent',
-        },
-      ],
-    },
-  ];
-
   it('should get the number of inactive workplaces', async () => {
     sinon.stub(findInactiveWorkplaces, 'findInactiveWorkplaces').returns(dummyInactiveWorkplaces);
-    sinon.stub(findParentWorkplaces, 'findParentWorkplaces').returns(dummyParentWorkplaces);
 
     const req = httpMocks.createRequest({
       method: 'GET',
@@ -99,14 +57,12 @@ describe('server/routes/admin/email-campaigns/inactive-workplaces', () => {
     await inactiveWorkplaceRoutes.getInactiveWorkplaces(req, res);
     const response = res._getJSONData();
 
-    expect(response.inactiveWorkplaces).to.deep.equal(3);
+    expect(response.inactiveWorkplaces).to.deep.equal(2);
   });
 
   describe('createCampaign', async () => {
     it('should create a campaign', async () => {
       sinon.stub(findInactiveWorkplaces, 'findInactiveWorkplaces').returns(dummyInactiveWorkplaces);
-      sinon.stub(findParentWorkplaces, 'findParentWorkplaces').returns(dummyParentWorkplaces);
-
       const sendEmailMock = sinon.stub(sendEmail, 'sendEmail').returns();
       const userMock = sinon.stub(models.user, 'findByUUID').returns({
         id: 1,
@@ -133,7 +89,7 @@ describe('server/routes/admin/email-campaigns/inactive-workplaces', () => {
 
       expect(response).to.deep.equal({
         date: '2021-01-01',
-        emails: 3,
+        emails: 2,
       });
 
       sinon.assert.calledOnce(createEmailCampaignHistoryMock);
@@ -144,7 +100,6 @@ describe('server/routes/admin/email-campaigns/inactive-workplaces', () => {
       });
       sinon.assert.calledWith(sendEmailMock, dummyInactiveWorkplaces[0]);
       sinon.assert.calledWith(sendEmailMock, dummyInactiveWorkplaces[1]);
-      sinon.assert.calledWith(sendEmailMock, dummyParentWorkplaces[0]);
     });
 
     it('should get the email campaign history', async () => {
